// src/modules/license/license.enums.ts
export enum LicenseValidationCode {
	LICENSE_NOT_FOUND = 'LICENSE_NOT_FOUND',
	LICENSE_DELETED = 'LICENSE_DELETED',
	LICENSE_INACTIVE = 'LICENSE_INACTIVE',
	LICENSE_ORGANIZATION_INACTIVE = 'LICENSE_ORGANIZATION_INACTIVE',
	LICENSE_EXPIRED = 'LICENSE_EXPIRED',
	LICENSE_SEATS_EXCEEDED = 'LICENSE_SEATS_EXCEEDED',
	LICENSE_SERVER_INACCESSIBLE = 'LICENSE_SERVER_INACCESSIBLE',
	LICENSE_VALIDATED = 'LICENSE_VALIDATED',
}

export enum SeatAllocationCode {
	LICENSE_NOT_FOUND = 'LICENSE_NOT_FOUND',
	LICENSE_DELETED = 'LICENSE_DELETED',
	LICENSE_INACTIVE = 'LICENSE_INACTIVE',
	LICENSE_EXPIRED = 'LICENSE_EXPIRED',
	LICENSE_SEATS_EXCEEDED = 'LICENSE_SEATS_EXCEEDED',
	SEAT_USER_NOT_FOUND = 'SEAT_USER_NOT_FOUND',
	SEAT_ALREADY_EXISTS = 'SEAT_ALREADY_EXISTS',
	SEAT_ALLOCATED = 'SEAT_ALLOCATED',
}